refactor(passport): avoid default export and rename passport to index
